requirements.txt
aenum==3.1.15
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
confection==0.1.3
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.2
fastapi==0.104.1
fiona==1.9.5
fonttools==4.44.0
geographiclib==2.0
geopandas==0.14.1
geopy==2.4.0
h11==0.14.0
httptools==0.6.1
httpx==0.27.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.5
langcodes==3.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.1
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
preshed==3.0.9
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sentinelhub==3.9.3
shapely==2.0.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.27.0
thinc==8.2.1
tifffile==2023.9.26
tomli==2.0.1
tomli_w==1.0.0
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
utm==0.7.0
uvicorn==0.24.0.post1
uvloop==0.19.0
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
